+++ Documentation/NuttX.html
@@ -6746,7 +6746,7 @@ BFD_ASSERT (*plt_offset != (bfd_vma) -1);
 </pre></ul>
       <p>
         No workaround is known at this time.  This is a show stopper for M16C.
-        Refer to the NuttX board <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/renesas/m32262f8/skp16c26/README.txt" target="_blank">README</a> file for further information.
+        Refer to the NuttX board <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/renesas/m16c/skp16c26/README.txt" target="_blank">README</a> file for further information.
       </p>
     </ul>
   </td>
