+++ Documentation/README.html
@@ -440,9 +440,9 @@ nuttx/
  |   |       `- or1k/
  |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/or1k/mor1kx/or1k/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |- renesas/
- |   |   |- m32262f8/
+ |   |   |- m16c/
  |   |   |   `- skp16c26/
- |   |   |       `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/renesas/m32262f8/skp16c26/README.txt" target="_blank"><b><i>README.txt</i></b></a>
+ |   |   |       `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/renesas/m16c/skp16c26/README.txt" target="_blank"><b><i>README.txt</i></b></a>
  |   |   `-sh7032/
  |   |       `- us7032evb1/
  |   |           `- <a href="https://bitbucket.org/nuttx/nuttx/src/master/boards/renesas/sh7032/us7032evb1/README.txt" target="_blank"><b><i>README.txt</i></b></a>
