+++ boards/renesas/m16c/skp16c26/Kconfig
