+++ boards/renesas/m16c/skp16c26/configs/ostest/defconfig
@@ -9,7 +9,7 @@
 CONFIG_ARCH="renesas"
 CONFIG_ARCH_BOARD="skp16c26"
 CONFIG_ARCH_BOARD_SKP16C26=y
-CONFIG_ARCH_CHIP="m32262f8"
+CONFIG_ARCH_CHIP="m16c"
 CONFIG_ARCH_CHIP_M30262F8=y
 CONFIG_ARCH_INTERRUPTSTACK=128
 CONFIG_ARCH_RENESAS=y
