+++ boards/renesas/m16c/skp16c26/scripts/skp16c26.ld
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/renesas/m32262f8/skp16c26/scripts/skp16c26.ld
+ * boards/renesas/m16c/skp16c26/scripts/skp16c26.ld
  *
  *   Copyright (C) 2009, 2017 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
