+++ boards/renesas/m16c/skp16c26/src/skp16c26.h
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/renesas/m32262f8/skp16c26/src/scp16c26.h
+ * boards/renesas/m16c/skp16c26/src/scp16c26.h
  *
  *   Copyright (C) 2009 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
